Added baseline search for existing recordings
This change speeds up the generate_recording_dir by a lot. It was taking
~10ms per existing recording dir (taking ~30s just to start a recording
when 3000 directories exist on the SD card). There is no noticable delay
with this new method. It also correctly tracks the earliest recording
even if it's not slot 0, and is able to prune directories properly if
the SD card dropoff config is enabled.